backend/pyproject.toml
[project]
name = "osmo-command-center"
version = "0.1.0"
description = "Voice-driven Command Center backend — Phase 1: Calendar"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "pydantic[email]>=2.0",
    "pydantic-settings",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "redis",
    "openai",
    "google-api-python-client",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "httpx",
    "cryptography",
    "python-jose[cryptography]",
    "uvicorn",
    "structlog",
    "sentry-sdk[fastapi]",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "httpx",
    "ruff",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
